fix：应用层整体代码优化和功能完善

+++ app/src/main/java/com/yk/memo/ui/main/MainActivity.java
@@ -1,18 +1,19 @@
 package com.yk.memo.ui.main;
 
 import android.os.Bundle;
+import android.text.TextUtils;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.appcompat.widget.PopupMenu;
 import androidx.appcompat.widget.SearchView;
 import androidx.appcompat.widget.Toolbar;
+import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import com.yk.base.eventposter.EventPoster;
@@ -22,6 +23,7 @@ import com.yk.memo.R;
 import com.yk.memo.data.adapter.NoteAdapter;
 import com.yk.memo.data.bean.Note;
 import com.yk.memo.data.event.NoteAddEvent;
+import com.yk.memo.data.event.NoteRemoveEvent;
 import com.yk.memo.data.event.NoteUpdateEvent;
 import com.yk.memo.ui.edit.EditActivity;
 
@@ -29,23 +31,18 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class MainActivity extends BaseMvpActivity<IMainView, MainPresenter> implements IMainView {
-    private static final String TAG = "MainActivity";
-
     private Toolbar toolbar;
     private SearchView searchView;
     private SwipeRefreshLayout swipeRefreshLayout;
     private RecyclerView rvNote;
 
-    private final List<Note> noteList = new ArrayList<>();
     private NoteAdapter noteAdapter;
 
-    private MenuItem deleteMenuItem;
-
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
         EventPoster.getInstance().register(this);
+        setContentView(R.layout.activity_main);
         findView();
         initData();
         bindEvent();
@@ -72,13 +69,53 @@ public class MainActivity extends BaseMvpActivity<IMainView, MainPresenter> impl
         loadAllNote();
     }
 
+    private void bindEvent() {
+        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                return false;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String newText) {
+                noteAdapter.getFilter().filter(newText);
+                return true;
+            }
+        });
+
+        swipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
+            @Override
+            public void onRefresh() {
+                loadAllNote();
+            }
+        });
+
+        noteAdapter.setOnItemListener(new NoteAdapter.OnItemListener() {
+            @Override
+            public void onItemClick(View view, Note note) {
+                onNoteItemClick(note);
+            }
+
+            @Override
+            public void onItemLongClick(View view, Note note) {
+                onNoteItemLongClick(note);
+            }
+
+            @Override
+            public void onItemMoreClick(View view, Note note) {
+                onNoteItemMoreClick(view, note);
+            }
+        });
+    }
+
     private void initToolbar() {
         setSupportActionBar(toolbar);
     }
 
     private void initRvNote() {
-        StaggeredGridLayoutManager layoutManager = new StaggeredGridLayoutManager(2, StaggeredGridLayoutManager.VERTICAL);
-        noteAdapter = new NoteAdapter(noteList);
+        LinearLayoutManager layoutManager = new LinearLayoutManager(this);
+        layoutManager.setOrientation(LinearLayoutManager.VERTICAL);
+        noteAdapter = new NoteAdapter(new ArrayList<>());
         rvNote.setLayoutManager(layoutManager);
         rvNote.setAdapter(noteAdapter);
     }
@@ -94,80 +131,36 @@ public class MainActivity extends BaseMvpActivity<IMainView, MainPresenter> impl
         presenter.loadAllNote();
     }
 
-    private void bindEvent() {
-        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-            @Override
-            public boolean onQueryTextSubmit(String query) {
-                return false;
+    private void onNoteItemClick(Note note) {
+        if (noteAdapter.isMoreSelectMode()) {
+            noteAdapter.selectNote(note);
+            if (noteAdapter.getSelectNoteList().isEmpty()) {
+                // 如果没有选中的note，则退出多选模式
+                noteAdapter.setMoreSelectMode(false);
+                hideDeleteMenuItem();
             }
-
-            @Override
-            public boolean onQueryTextChange(String newText) {
-                Log.d(TAG, "onQueryTextChange: " + newText);
-                noteAdapter.getFilter().filter(newText);
-                return true;
-            }
-        });
-
-        swipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
-            @Override
-            public void onRefresh() {
-                presenter.loadAllNote();
-            }
-        });
-
-        noteAdapter.setOnItemListener(new NoteAdapter.OnItemListener() {
-            @Override
-            public void onItemClick(View view, Note note) {
-                if (noteAdapter.isMoreSelectMode()) {
-                    noteAdapter.selectNote(note);
-                } else {
-                    startEditActivity(note);
-                }
-            }
-
-            @Override
-            public void onItemLongClick(View view, Note note) {
-                if (noteAdapter.isMoreSelectMode()) {
-                    return;
-                }
-                noteAdapter.setMoreSelectMode(true);
-                noteAdapter.selectNote(note);
-                showDeleteMenu();
-            }
-
-            @Override
-            public void onItemMoreClick(View view, Note note) {
-                showItemNotePopupMenu(view, note);
-            }
-        });
+        } else {
+            startEditActivity(note);
+        }
     }
 
-    private void startEditActivity(Note note) {
-        EditActivity.startEditActivity(this, note != null ? note.getId() : -1);
-    }
-
-    private void deleteBatch() {
-        List<Note> noteList = noteAdapter.getSelectNoteList();
-        if (noteList.isEmpty()) {
-            Toast.makeText(this, "至少选中一个", Toast.LENGTH_SHORT).show();
+    private void onNoteItemLongClick(Note note) {
+        if (noteAdapter.isMoreSelectMode()) {
             return;
         }
-        presenter.deleteBatchNote(noteList);
+        noteAdapter.setMoreSelectMode(true);
+        noteAdapter.selectNote(note);
+        showDeleteMenuItem();
     }
 
-    private void showItemNotePopupMenu(View view, Note note) {
+    private void onNoteItemMoreClick(View view, Note note) {
         PopupMenu popupMenu = new PopupMenu(this, view);
         popupMenu.getMenuInflater().inflate(R.menu.menu_item_note, popupMenu.getMenu());
         popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem item) {
-                switch (item.getItemId()) {
-                    case R.id.menu_item_note_delete:
-                        presenter.deleteNote(note);
-                        break;
-                    default:
-                        break;
+                if (item.getItemId() == R.id.menu_item_note_delete) {
+                    presenter.deleteNote(note);
                 }
                 return true;
             }
@@ -175,14 +168,24 @@ public class MainActivity extends BaseMvpActivity<IMainView, MainPresenter> impl
         popupMenu.show();
     }
 
-    private void showDeleteMenu() {
+    private void startEditActivity(Note note) {
+        EditActivity.start(this, note);
+    }
+
+    private void deleteNoteList() {
+        presenter.deleteNoteList(noteAdapter.getSelectNoteList());
+    }
+
+    private MenuItem deleteMenuItem;
+
+    private void showDeleteMenuItem() {
         if (deleteMenuItem == null) {
             return;
         }
         deleteMenuItem.setVisible(true);
     }
 
-    private void hideDeleteMenu() {
+    private void hideDeleteMenuItem() {
         if (deleteMenuItem == null) {
             return;
         }
@@ -190,82 +193,19 @@ public class MainActivity extends BaseMvpActivity<IMainView, MainPresenter> impl
     }
 
     @Override
-    public MainPresenter createPresenter() {
-        return new MainPresenter();
-    }
-
-    @Subscribe(threadMode = Subscribe.Thread.UI)
-    public void onNoteUpdate(NoteUpdateEvent event) {
-        if (event == null) {
-            return;
-        }
-
-        Note note = noteAdapter.findNoteForId(event.getNoteId());
-
-        if (note == null) {
-            return;
-        }
-
-        note.setSrc(event.getSrc());
-        note.setUpdateTime(event.getUpdateTime());
-        note.setSpanStrBuilder(event.getSpanStrBuilder());
-
-        noteAdapter.topNote(note);
-    }
-
-    @Subscribe(threadMode = Subscribe.Thread.UI)
-    public void onNoteAdd(NoteAddEvent event) {
-        if (event == null) {
-            return;
-        }
-
-        noteAdapter.refreshDataAdd(event.getNote(), true);
-
-        noteAdapter.getFilter().filter(searchView.getQuery());
-    }
-
-    @Override
-    public void onLoadNoteList(List<Note> noteList) {
-        noteAdapter.refreshDataAdd(noteList);
-        swipeRefreshLayout.setRefreshing(false);
-
-        noteAdapter.getFilter().filter(searchView.getQuery());
-    }
-
-    @Override
-    public void onDeleteNote(boolean success, Note note) {
-        Toast.makeText(this, success ? "delete success" : "delete fail", Toast.LENGTH_SHORT).show();
-        if (!success) {
-            return;
-        }
-        noteAdapter.refreshDataRemove(note);
-    }
-
-    @Override
-    public void onDeleteNoteList(boolean success, List<Note> noteList) {
-        Toast.makeText(this, success ? "delete list success" : "delete fail", Toast.LENGTH_SHORT).show();
-        if (!success) {
-            return;
-        }
-        noteAdapter.refreshDataRemove(noteList);
-        noteAdapter.setMoreSelectMode(false);
-        hideDeleteMenu();
-    }
-
-    @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.menu_main, menu);
         deleteMenuItem = menu.findItem(R.id.menu_main_delete);
-        hideDeleteMenu();
+        hideDeleteMenuItem();
         return true;
     }
 
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        if (item.getItemId() == R.id.menu_main_edit) {
+        if (item.getItemId() == R.id.menu_main_delete) {
+            deleteNoteList();
+        } else if (item.getItemId() == R.id.menu_main_edit) {
             startEditActivity(null);
-        } else if (item.getItemId() == R.id.menu_main_delete) {
-            deleteBatch();
         }
         return true;
     }
@@ -274,9 +214,95 @@ public class MainActivity extends BaseMvpActivity<IMainView, MainPresenter> impl
     public void onBackPressed() {
         if (noteAdapter.isMoreSelectMode()) {
             noteAdapter.setMoreSelectMode(false);
-            hideDeleteMenu();
+            hideDeleteMenuItem();
             return;
         }
         super.onBackPressed();
+    }
+
+    @Override
+    public MainPresenter createPresenter() {
+        return new MainPresenter();
+    }
+
+    @Subscribe(threadMode = Subscribe.Thread.UI)
+    public void onNoteAddEvent(NoteAddEvent event) {
+        if (event == null) {
+            return;
+        }
+        noteAdapter.refreshDataAdd(event.getNote(), event.isNeedTop());
+
+        if (event.isNeedTop()) {
+            rvNote.scrollToPosition(0);
+        }
+    }
+
+    @Subscribe(threadMode = Subscribe.Thread.UI)
+    public void onNoteUpdateEvent(NoteUpdateEvent event) {
+        Log.d(TAG, "onNoteUpdateEvent: " + event);
+        if (event == null) {
+            return;
+        }
+        Note eventNote = event.getNote();
+
+        Note note = noteAdapter.findNoteForId(eventNote.getId());
+        note.setSrc(eventNote.getSrc());
+        note.setUpdateTime(eventNote.getUpdateTime());
+
+        if (event.isNeedTop()) {
+            noteAdapter.topNote(note);
+            rvNote.scrollToPosition(0);
+        }
+    }
+
+    @Subscribe(threadMode = Subscribe.Thread.UI)
+    public void onNoteRemoveEvent(NoteRemoveEvent event) {
+        if (event == null) {
+            return;
+        }
+        Note eventNote = event.getNote();
+
+        Note note = noteAdapter.findNoteForId(eventNote.getId());
+        noteAdapter.refreshDataRemove(note);
+    }
+
+    private static final String TAG = "MainActivity2";
+
+    @Override
+    public void onLoadNoteList(List<Note> noteList) {
+        noteAdapter.refreshDataAdd(noteList);
+        swipeRefreshLayout.setRefreshing(false);
+
+        String query = searchView.getQuery().toString();
+        if (!TextUtils.isEmpty(query)) {
+            noteAdapter.getFilter().filter(query);
+        }
+    }
+
+    @Override
+    public void onDeleteNote(Note note) {
+        noteAdapter.refreshDataRemove(note);
+    }
+
+    @Override
+    public void onDeleteNoteList(List<Note> noteList) {
+        noteAdapter.refreshDataRemove(noteList);
+        noteAdapter.setMoreSelectMode(false);
+        hideDeleteMenuItem();
+    }
+
+    @Override
+    public void onLoadNoteListError(Exception e) {
+
+    }
+
+    @Override
+    public void onDeleteNoteError(Exception e) {
+
+    }
+
+    @Override
+    public void onDeleteNoteListError(Exception e) {
+
     }
 }
